coding/Chapter_2_Routing/requirements.txt
# Chapter 2 Routing dependencies
#langchain>=0.2.0
#langchain-community>=0.2.0
#langchain-openai>=0.1.0
#langgraph>=0.1.0
python-dotenv>=1.0.0
numpy>=1.24.0
